pyproject.toml
[tool.poetry]
name = "apis-rdf-devops"
version = "0.1.0"
description = "Generic APIS project for auto devops"
authors = ["Matthias Schlögl <m.schloegl@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "~3.8"
whitenoise = "^5.2.0"
sentry-sdk = "^0.19.5"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
acdh-django-charts = "^0.5.3"
django-allow-cidr = "^0.3.1"
mysqlclient = "^2.0.3"
django-extensions = "^3.1.3"
django-csp = "^3.7"
apis-bibsonomy = "~0.3"
apis-highlighter = "~0.9"
viecpro-vis = "^0.1.10"
apis-core = "~0.17"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.0"
pylint = "^2.6.0"
mistune = "<2.0.0"
dephell = {extras = ["full"], version = "^0.8.3"}
jupyter = "^1.0.0"
jupyter-tabnine = "^1.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
